requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sentence-transformers==2.2.2
faiss-cpu==1.9.0
openai==0.27.8
numpy==1.26.4
pydantic==1.10.12
